changed dimension variables to size_t, fixed warnings

+++ src/plugins/intel_gna/src/gna_plugin.cpp
@@ -1066,9 +1066,9 @@ RequestStatus GNAPlugin::WaitFor(uint32_t request_idx, int64_t millisTimeout) {
         auto is1D = output_layout == InferenceEngine::Layout::C;
         auto isScalar = output_layout == InferenceEngine::Layout::SCALAR;
         auto is3D = output_layout == InferenceEngine::Layout::CHW;
-        uint32_t batchSize = static_cast<uint32_t>((is1D || isScalar || is3D) ? 1 : dims[0]);
-        uint32_t elementsPerBatch = static_cast<uint32_t>(
-            isScalar ? 1 : (is1D ? dims.front() : details::product(++std::begin(dims), std::end(dims))));
+        size_t batchSize = (is1D || isScalar || is3D) ? 1 : dims[0];
+        size_t elementsPerBatch =
+            isScalar ? 1 : (is1D ? dims.front() : details::product(++std::begin(dims), std::end(dims)));
 
         OutputDesc& gna_output_desc = outputs_.at(output_name);
         Blob::Ptr gna_output_blob = nullptr;
